Read configuration from rc-filens and shell environment The configuration 'object' is initialised with hardcoded defaults. Then, in several passes, new values are added and already present values replaced, with configuration values from different sources. In order:
 1. /etc/teve.conf (*)
 2. $HOME/.teverc (*) (**)
 3. Shell environment variables

And, since the command line parameters are parsed after config.scm is loaded
and the configuration 'object' is defined:

 4. Command line parameters.

(*) The configuration files uses a standard INI-file format:

    [operators]
    default=download

The above will result in (*cfg* 'operators 'default) returning
'download.

(**) The path and filename of this file can be changed from the default
$HOME/.teverc by setting the shell environment variable TEVE_RC

This fixes issue #16.