Use file-exists? instead of invoking where.exe on Windows Both process and process-run from the posix egg never finishes.
This limits the external program detection somewhat on Windows. If the specified (or default) name of the program isn't absolute (or relative in a way that happens to work in the current working directory), it won't get detected in Windows.

A proper way to find available external programs would require any relative paths to be checked against each path in the PATH environment variable.

The recommended solution is to simply use absolute paths to external programs.

This fixes #32.